requirements.txt
streamlit
streamlit-option-menu
pandas
requests
plotly
joblib
scikit-learn
python-dotenv